requirements.txt
tensorflow
Pillow
numpy
Flask
# Specify version if needed.